backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
qdrant-client>=1.12.0
openai>=1.55.0
docling>=2.14.0
semantic-router>=0.0.46
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
ragas>=0.2.0
datasets>=3.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9
httpx>=0.27.0
PyJWT>=2.9.0
passlib[bcrypt]>=1.7.4
pytest>=8.0.0
tiktoken>=0.8.0
